Refactor: Add threshold logic to review card dto
